requirements.txt
python-dotenv~=1.0.0
PyGithub~=2.2.0